+++ src/datakit/ml/__init__.py
@@ -1,11 +1,12 @@
 """Machine Learning subpackage for DataKit."""
 from datakit.ml.ml_helpers import check_imbalance, create_cv_splits, encode_target_labels
-from datakit.ml.models import fit_model
+from datakit.ml.models import fit_model, tune_model
 from datakit.ml.prepare import prepare_data
 
 __all__ = [
     "prepare_data",
     "fit_model",
+    "tune_model",
     "encode_target_labels",
     "check_imbalance",
     "create_cv_splits",